DocBot/requirements.txt
pandas
numpy
PyPDF2
langchain
tiktoken
python-dotenv
pinecone-client
torch
streamlit
openai
flask
textract
